[스터디 목록] 코드 리팩토링: 패키지 경로 수정 및 불필요한 항목 제거
