+++ .idea/gradle.xml
@@ -11,7 +11,6 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/feature-studylist" />
           </set>
         </option>
       </GradleProjectSettings>
