+++ app/src/main/java/com/example/smartee/navigation/NavGraph.kt
@@ -15,6 +15,7 @@ import com.example.smartee.ui.register.RegisterScreen
 import com.example.smartee.ui.study.StudyCreateScreen
 import com.example.smartee.ui.study.StudyListScreen
 
+
 @Composable
 fun SmarteeNavGraph(navController: NavHostController) {
     NavHost(navController, startDestination = Screen.Login.route) {
@@ -29,9 +30,6 @@ fun SmarteeNavGraph(navController: NavHostController) {
         }
         composable(Screen.StudyCreate.route) {
             StudyCreateScreen(navController)
-        }
-        composable(Screen.StudyList.route) {
-            StudyListScreen(navController)
         }
 
         //스터디 목록 관련
