+++ app/src/main/java/com/example/smartee/ui/study/studyList/topbar/StudyListTopBar.kt
@@ -11,8 +11,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.feature_studylist.uicomponents.topbar.address.AddressList
-import com.example.feature_studylist.viewmodel.StudyViewModel
+import com.example.smartee.ui.study.studyList.topbar.address.AddressList
+import com.example.smartee.viewmodel.StudyViewModel
 
 @Composable
 fun StudyListTopBar(
