+++ app/src/main/java/com/example/smartee/ui/study/studyList/topbar/address/AddressList.kt
@@ -10,7 +10,7 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.feature_studylist.viewmodel.StudyViewModel
+import com.example.smartee.viewmodel.StudyViewModel
 
 @Composable
 fun AddressList(
