--- feature-studylist/src/main/java/com/example/feature_studylist/model/factory/AddressListFactory.kt
@@ -1,5 +1,0 @@
-package com.example.feature_studylist.model.factory
-
-object AddressListFactory {
-    fun makeAddressList() = mutableListOf("", "군자동", "구의제3동", "휘경동")
-}
